Refactor to work with the v0.2.0 feedstock references. (#2)
This handles the new representation of genes, where the concept of a single
primary identifier no longer exists. Rather, all identifiers are treated
equally, allowing callers to pull down the exact identifiers of interest.

In addition, all gesel functions are now species-specific. This reduces the
payload that needs to be transferred to the client when only a single
species is of interest (which is almost always the case).