Pass a null value to trigger initialization without further requests.
This is useful for concurrent requests to some functions like
fetchSetsByGene, where the transparent initialization of various globals
is async and unpredictable. As a consequence, we run into race
conditions where the globals are not fully initialized, causing other
calls to repeat the initialization unnecessarily, or proceed to the
request (and fail). We can't really apply mutexes here as there aren't
actually multiple threads so deadlocking on the waiting task is a
possibility; and we can't easily predict the behavior of the task
scheduler; so the best approach is to ask callers to force
initialization via a null input if they anticipate multiple later calls.