Compute p-values directly from findOverlappingSets.
This makes life easier and also ensures that we give it the right arguments
with respect to only considering genes that are involved in at least one set.
Specifically, we want to ensure that useless genes are not counted in the list
size, consistent with the use of effectiveNumberOfGenes for the universe.